refactored items API
